Use tree-diff for config roundtrip tests.
