+++ cabal-install/Distribution/Client/IndexUtils/Timestamp.hs
@@ -39,7 +39,7 @@ import           GHC.Generics (Generic)
 
 -- | UNIX timestamp (expressed in seconds since unix epoch, i.e. 1970).
 newtype Timestamp = TS Int64 -- Tar.EpochTime
-                  deriving (Eq,Ord,Enum,NFData,Show)
+                  deriving (Eq,Ord,Enum,NFData,Show,Generic)
 
 epochTimeToTimestamp :: Tar.EpochTime -> Maybe Timestamp
 epochTimeToTimestamp et
