+++ cabal-install/tests/UnitTests/Distribution/Client/GenericInstances.hs
@@ -1,0 +1,10 @@
+{-# LANGUAGE DeriveGeneric        #-}
+{-# LANGUAGE StandaloneDeriving   #-}
+{-# OPTIONS_GHC -fno-warn-orphans #-}
+
+module UnitTests.Distribution.Client.GenericInstances () where
+
+import Network.URI
+import GHC.Generics
+
+deriving instance Generic URIAuth
